move suggestion handlers to better directory structure
