+++ src/backend/web/handlers/suggestions/tests/suggest_event_media_controller_test.py
@@ -6,22 +6,26 @@ from bs4 import BeautifulSoup
 from google.cloud import ndb
 from werkzeug.test import Client
 
+from backend.common.consts.event_type import EventType
 from backend.common.consts.media_type import MediaType
 from backend.common.consts.suggestion_state import SuggestionState
+from backend.common.models.event import Event
 from backend.common.models.suggestion import Suggestion
 from backend.common.models.suggestion_dict import SuggestionDict
-from backend.common.models.team import Team
-from backend.web.handlers.tests.conftest import CapturedTemplate
+from backend.web.handlers.conftest import CapturedTemplate
 
 
 @pytest.fixture(autouse=True)
-def storeTeam(ndb_client: ndb.Client) -> None:
+def storeEvent(ndb_client: ndb.Client) -> None:
     with ndb_client.context():
-        team = Team(
-            id="frc1124",
-            team_number=1124,
+        event = Event(
+            id="2016nyny",
+            event_type_enum=EventType.REGIONAL,
+            name="NYC",
+            event_short="NYC",
+            year=2016,
         )
-        team.put()
+        event.put()
 
 
 def assert_template_status(
@@ -29,34 +33,34 @@ def assert_template_status(
 ) -> None:
     template = captured_templates[0][0]
     context = captured_templates[0][1]
-    assert template.name == "suggestions/suggest_team_social_media.html"
+    assert template.name == "suggestions/suggest_event_media.html"
     assert context["status"] == status
 
 
 def test_login_redirect(web_client: Client) -> None:
-    response = web_client.get("/suggest/team/social_media?team_key=frc1124")
+    response = web_client.get("/suggest/event/media?event_key=2016nyny")
     assert response.status_code == 302
     assert urlparse(response.headers["Location"]).path == "/account/login"
 
 
-def test_get_no_team(login_user, web_client: Client) -> None:
-    response = web_client.get("/suggest/team/social_media")
+def test_get_no_event(login_user, web_client: Client) -> None:
+    response = web_client.get("/suggest/event/media")
     assert response.status_code == 404
 
 
-def test_get_bad_team(login_user, web_client: Client) -> None:
-    response = web_client.get("/suggest/team/social_media?team_key=frc254")
+def test_get_bad_event(login_user, web_client: Client) -> None:
+    response = web_client.get("/suggest/event/media?event_key=2016foo")
     assert response.status_code == 404
 
 
 def test_get_form(login_user, web_client: Client) -> None:
-    response = web_client.get("/suggest/team/social_media?team_key=frc1124&year=2016")
+    response = web_client.get("/suggest/event/media?event_key=2016nyny")
     assert response.status_code == 200
 
     soup = BeautifulSoup(response.data, "html.parser")
-    form = soup.find("form", id="suggest_social_media")
+    form = soup.find("form", id="suggest_media")
     assert form is not None
-    assert form["action"] == "/suggest/team/social_media"
+    assert form["action"] == "/suggest/event/media"
     assert form["method"] == "post"
 
     csrf = form.find(attrs={"name": "csrf_token"})
@@ -64,16 +68,16 @@ def test_get_form(login_user, web_client: Client) -> None:
     assert csrf["type"] == "hidden"
     assert csrf["value"] is not None
 
-    team_key = form.find(attrs={"name": "team_key"})
-    assert team_key is not None
-    assert team_key["type"] == "hidden"
-    assert team_key["value"] == "frc1124"
+    event_key = form.find(attrs={"name": "event_key"})
+    assert event_key is not None
+    assert event_key["type"] == "hidden"
+    assert event_key["value"] == "2016nyny"
 
     assert form.find(attrs={"name": "media_url"}) is not None
 
 
-def test_submit_no_team(login_user, ndb_client: ndb.Client, web_client: Client) -> None:
-    resp = web_client.post("/suggest/team/social_media", data={})
+def test_submit_no_data(login_user, ndb_client: ndb.Client, web_client: Client) -> None:
+    resp = web_client.post("/suggest/event/media", data={}, follow_redirects=True)
     assert resp.status_code == 404
 
     # Assert no suggestions were written
@@ -81,15 +85,13 @@ def test_submit_no_team(login_user, ndb_client: ndb.Client, web_client: Client) 
         assert Suggestion.query().fetch() == []
 
 
-def test_submit_bad_team(
+def test_submit_bad_event(
     login_user, ndb_client: ndb.Client, web_client: Client
 ) -> None:
-    response = web_client.post(
-        "/suggest/team/social_media",
-        data={"team_key": "frc254"},
-        follow_redirects=True,
+    resp = web_client.post(
+        "/suggest/event/media", data={"event_key": "2016foo"}, follow_redirects=True
     )
-    assert response.status_code == 404
+    assert resp.status_code == 404
 
     # Assert no suggestions were written
     with ndb_client.context():
@@ -103,8 +105,31 @@ def test_submit_empty_form(
     captured_templates: List[CapturedTemplate],
 ) -> None:
     resp = web_client.post(
-        "/suggest/team/social_media",
-        data={"team_key": "frc1124", "media_url": ""},
+        "/suggest/event/media",
+        data={"event_key": "2016nyny"},
+        follow_redirects=True,
+    )
+    assert resp.status_code == 200
+    assert_template_status(captured_templates, "bad_url")
+
+    # Assert the correct dialog shows
+    soup = BeautifulSoup(resp.data, "html.parser")
+    assert soup.find(id="bad_url-alert") is not None
+
+    # Assert no suggestions were written
+    with ndb_client.context():
+        assert Suggestion.query().fetch() == []
+
+
+def test_submit_bad_url(
+    login_user,
+    ndb_client: ndb.Client,
+    web_client: Client,
+    captured_templates: List[CapturedTemplate],
+) -> None:
+    resp = web_client.post(
+        "/suggest/event/media",
+        data={"event_key": "2016nyny", "media_url": "asdf"},
         follow_redirects=True,
     )
     assert resp.status_code == 200
@@ -126,10 +151,10 @@ def test_suggest_media(
     captured_templates: List[CapturedTemplate],
 ) -> None:
     resp = web_client.post(
-        "/suggest/team/social_media",
+        "/suggest/event/media",
         data={
-            "team_key": "frc1124",
-            "media_url": "https://github.com/frc1124",
+            "event_key": "2016nyny",
+            "media_url": "https://www.youtube.com/watch?v=H-54KMwMKY0",
         },
         follow_redirects=True,
     )
@@ -145,15 +170,13 @@ def test_suggest_media(
         suggestion = cast(Suggestion, Suggestion.query().fetch()[0])
         assert suggestion is not None
         assert suggestion.review_state == SuggestionState.REVIEW_PENDING
-        assert suggestion.target_key == "frc1124"
+        assert suggestion.target_model == "event_media"
         assert suggestion.contents == SuggestionDict(
-            year=None,
-            reference_type="team",
-            reference_key="frc1124",
-            foreign_key="frc1124",
-            profile_url="https://github.com/frc1124",
-            is_social=True,
-            media_type_enum=MediaType.GITHUB_PROFILE,
-            default_preferred=False,
-            site_name="GitHub Profile",
+            year=2016,
+            reference_type="event",
+            reference_key="2016nyny",
+            foreign_key="H-54KMwMKY0",
+            is_social=False,
+            media_type_enum=MediaType.YOUTUBE_VIDEO,
+            site_name="YouTube Video",
         )
