+++ src/backend/web/handlers/suggestions/tests/suggest_team_media_controller_test.py
@@ -6,26 +6,22 @@ from bs4 import BeautifulSoup
 from google.cloud import ndb
 from werkzeug.test import Client
 
-from backend.common.consts.event_type import EventType
 from backend.common.consts.media_type import MediaType
 from backend.common.consts.suggestion_state import SuggestionState
-from backend.common.models.event import Event
 from backend.common.models.suggestion import Suggestion
 from backend.common.models.suggestion_dict import SuggestionDict
-from backend.web.handlers.tests.conftest import CapturedTemplate
+from backend.common.models.team import Team
+from backend.web.handlers.conftest import CapturedTemplate
 
 
 @pytest.fixture(autouse=True)
-def storeEvent(ndb_client: ndb.Client) -> None:
+def storeTeam(ndb_client: ndb.Client):
     with ndb_client.context():
-        event = Event(
-            id="2016nyny",
-            event_type_enum=EventType.REGIONAL,
-            name="NYC",
-            event_short="NYC",
-            year=2016,
+        team = Team(
+            id="frc1124",
+            team_number=1124,
         )
-        event.put()
+        team.put()
 
 
 def assert_template_status(
@@ -33,34 +29,39 @@ def assert_template_status(
 ) -> None:
     template = captured_templates[0][0]
     context = captured_templates[0][1]
-    assert template.name == "suggestions/suggest_event_media.html"
+    assert template.name == "suggestions/suggest_team_media.html"
     assert context["status"] == status
 
 
 def test_login_redirect(web_client: Client) -> None:
-    response = web_client.get("/suggest/event/media?event_key=2016nyny")
+    response = web_client.get("/suggest/team/media?team_key=frc1124&year=2016")
     assert response.status_code == 302
     assert urlparse(response.headers["Location"]).path == "/account/login"
 
 
-def test_get_no_event(login_user, web_client: Client) -> None:
-    response = web_client.get("/suggest/event/media")
+def test_get_no_team(login_user, web_client: Client) -> None:
+    response = web_client.get("/suggest/team/media")
     assert response.status_code == 404
 
 
-def test_get_bad_event(login_user, web_client: Client) -> None:
-    response = web_client.get("/suggest/event/media?event_key=2016foo")
+def test_get_bad_team(login_user, web_client: Client) -> None:
+    response = web_client.get("/suggest/team/media?team_key=frc254&year=2016")
+    assert response.status_code == 404
+
+
+def test_get_bad_year(login_user, web_client: Client) -> None:
+    response = web_client.get("/suggest/team/media?team_key=frc1124&year=asdf")
     assert response.status_code == 404
 
 
 def test_get_form(login_user, web_client: Client) -> None:
-    response = web_client.get("/suggest/event/media?event_key=2016nyny")
+    response = web_client.get("/suggest/team/media?team_key=frc1124&year=2016")
     assert response.status_code == 200
 
     soup = BeautifulSoup(response.data, "html.parser")
     form = soup.find("form", id="suggest_media")
     assert form is not None
-    assert form["action"] == "/suggest/event/media"
+    assert form["action"] == "/suggest/team/media"
     assert form["method"] == "post"
 
     csrf = form.find(attrs={"name": "csrf_token"})
@@ -68,16 +69,21 @@ def test_get_form(login_user, web_client: Client) -> None:
     assert csrf["type"] == "hidden"
     assert csrf["value"] is not None
 
-    event_key = form.find(attrs={"name": "event_key"})
-    assert event_key is not None
-    assert event_key["type"] == "hidden"
-    assert event_key["value"] == "2016nyny"
+    team_key = form.find(attrs={"name": "team_key"})
+    assert team_key is not None
+    assert team_key["type"] == "hidden"
+    assert team_key["value"] == "frc1124"
+
+    year = form.find(attrs={"name": "year"})
+    assert year is not None
+    assert year["type"] == "hidden"
+    assert year["value"] == "2016"
 
     assert form.find(attrs={"name": "media_url"}) is not None
 
 
-def test_submit_no_data(login_user, ndb_client: ndb.Client, web_client: Client) -> None:
-    resp = web_client.post("/suggest/event/media", data={}, follow_redirects=True)
+def test_submit_no_team(login_user, ndb_client: ndb.Client, web_client: Client) -> None:
+    resp = web_client.post("/suggest/team/media", data={})
     assert resp.status_code == 404
 
     # Assert no suggestions were written
@@ -85,13 +91,15 @@ def test_submit_no_data(login_user, ndb_client: ndb.Client, web_client: Client) 
         assert Suggestion.query().fetch() == []
 
 
-def test_submit_bad_event(
+def test_submit_bad_team(
     login_user, ndb_client: ndb.Client, web_client: Client
 ) -> None:
-    resp = web_client.post(
-        "/suggest/event/media", data={"event_key": "2016foo"}, follow_redirects=True
+    response = web_client.post(
+        "/suggest/team/media",
+        data={"team_key": "frc254", "year": "2016"},
+        follow_redirects=True,
     )
-    assert resp.status_code == 404
+    assert response.status_code == 404
 
     # Assert no suggestions were written
     with ndb_client.context():
@@ -105,31 +113,8 @@ def test_submit_empty_form(
     captured_templates: List[CapturedTemplate],
 ) -> None:
     resp = web_client.post(
-        "/suggest/event/media",
-        data={"event_key": "2016nyny"},
-        follow_redirects=True,
-    )
-    assert resp.status_code == 200
-    assert_template_status(captured_templates, "bad_url")
-
-    # Assert the correct dialog shows
-    soup = BeautifulSoup(resp.data, "html.parser")
-    assert soup.find(id="bad_url-alert") is not None
-
-    # Assert no suggestions were written
-    with ndb_client.context():
-        assert Suggestion.query().fetch() == []
-
-
-def test_submit_bad_url(
-    login_user,
-    ndb_client: ndb.Client,
-    web_client: Client,
-    captured_templates: List[CapturedTemplate],
-) -> None:
-    resp = web_client.post(
-        "/suggest/event/media",
-        data={"event_key": "2016nyny", "media_url": "asdf"},
+        "/suggest/team/media",
+        data={"team_key": "frc1124", "year": "2016", "media_url": ""},
         follow_redirects=True,
     )
     assert resp.status_code == 200
@@ -151,10 +136,11 @@ def test_suggest_media(
     captured_templates: List[CapturedTemplate],
 ) -> None:
     resp = web_client.post(
-        "/suggest/event/media",
+        "/suggest/team/media",
         data={
-            "event_key": "2016nyny",
-            "media_url": "https://www.youtube.com/watch?v=H-54KMwMKY0",
+            "team_key": "frc1124",
+            "year": "2016",
+            "media_url": "http://imgur.com/aF8T5ZE",
         },
         follow_redirects=True,
     )
@@ -170,13 +156,14 @@ def test_suggest_media(
         suggestion = cast(Suggestion, Suggestion.query().fetch()[0])
         assert suggestion is not None
         assert suggestion.review_state == SuggestionState.REVIEW_PENDING
-        assert suggestion.target_model == "event_media"
+        assert suggestion.target_key == "frc1124"
         assert suggestion.contents == SuggestionDict(
             year=2016,
-            reference_type="event",
-            reference_key="2016nyny",
-            foreign_key="H-54KMwMKY0",
+            reference_type="team",
+            reference_key="frc1124",
+            foreign_key="aF8T5ZE",
             is_social=False,
-            media_type_enum=MediaType.YOUTUBE_VIDEO,
-            site_name="YouTube Video",
+            media_type_enum=MediaType.IMGUR,
+            default_preferred=False,
+            site_name="Imgur Image",
         )
